sd-watcher-afk/pyproject.toml
[tool.poetry]
name = "sd-watcher-afk"
version = "0.2.0"
description = "AFK watcher for ActivityWatch"
authors = ["Erik Bjäreholt <erik@bjareho.lt>"]
license = "MPL-2.0"

[tool.poetry.scripts]
aw-watcher-afk = "sd_watcher_afk:main"

[tool.poetry.dependencies]
python = "^3.8,<3.13"
aw-client = "^0.5"
pyobjc-framework-Quartz = { version = "*", platform = "darwin" }
pynput = "*"

# locked due to https://github.com/python-xlib/python-xlib/pull/242 leading to 100% CPU stalls
# see: https://github.com/ActivityWatch/sd-watcher-window/issues/89
python-xlib = { version = "0.31", platform = "linux" }

[tool.poetry.dev-dependencies]
mypy = "*"
macholib = {version = "*", platform = "darwin"}
pyinstaller = {version = "^5.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
